refactor: move client to its own file
